# server/internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for managing assignments, questions, classrooms and more.",
        "title": "ParableMind API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/classrooms": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Classrooms"
                ],
                "summary": "Get All Classrooms",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Classroom"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Classrooms"
                ],
                "summary": "Create Classroom",
                "parameters": [
                    {
                        "description": "Classroom JSON",
                        "name": "classroom",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Classroom"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Classroom"
                        }
                    }
                }
            }
        },
        "/classrooms/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Classrooms"
                ],
                "summary": "Get Classroom by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Classroom ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Classroom"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Classrooms"
                ],
                "summary": "Update Classroom",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Classroom ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Classroom",
                        "name": "classroom",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Classroom"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Classroom"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Classrooms"
                ],
                "summary": "Delete Classroom",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Classroom ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Classrooms"
                ],
                "summary": "Patch Classroom",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Classroom ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Partial updates",
                        "name": "updates",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/question-banks": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QuestionBanks"
                ],
                "summary": "Get All QuestionBanks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by topic",
                        "name": "topic",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by teacher ID",
                        "name": "teacherId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tags (comma-separated)",
                        "name": "tags",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.QuestionBank"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QuestionBanks"
                ],
                "summary": "Create QuestionBank",
                "parameters": [
                    {
                        "description": "QuestionBank JSON",
                        "name": "questionBank",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.QuestionBank"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.QuestionBank"
                        }
                    }
                }
            }
        },
        "/question-banks/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QuestionBanks"
                ],
                "summary": "Get QuestionBank by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "QuestionBank ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.QuestionBank"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QuestionBanks"
                ],
                "summary": "Update QuestionBank",
                "parameters": [
                    {
                        "type": "string",
                        "description": "QuestionBank ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated QuestionBank",
                        "name": "questionBank",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.QuestionBank"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.QuestionBank"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "QuestionBanks"
                ],
                "summary": "Delete QuestionBank",
                "parameters": [
                    {
                        "type": "string",
                        "description": "QuestionBank ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QuestionBanks"
                ],
                "summary": "Patch QuestionBank",
                "parameters": [
                    {
                        "type": "string",
                        "description": "QuestionBank ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updates",
                        "name": "updates",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Classroom": {
            "type": "object",
            "properties": {
                "assignmentIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "credits": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "tags": {
                    "description": "Added tags field",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "teacherIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "model.QuestionBank": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "tags": {
                    "description": "Added tags field",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "teacherId": {
                    "type": "string"
                },
                "topic": {
                    "type": "string"
                }
            }
        },
...
    }
}